+++ CHANGELOG.md
@@ -14,4 +14,5 @@ Put changelog here:
 - Added Mod Protocol API
   - A partial implementation of The Quilt Project's Registry Sync API.
 - Refactored Configured Features and Placement Modifiers
+- Replaced `frozenlib:client` entrypoint with an environment event
 - Removed embedded Mixin Extras in favor of Fabric Loader 0.15.0